♻ Refactor structure to be more hexagonal
Not sure if I should use an out adapter in my in adapter or if that
should go through a use case / service in my application layer but for
now it doesn't make sense to me to overcomplicate things. What I have
now is enough to test out the cucumber stuff and keep the bed the file
structure down in my head a bit so the concepts stick.